requirements.txt
python-dotenv>=1.0.1
pillow>=10.2.0
pandas>=2.2.0
openpyxl>=3.1.2
gspread>=5.12.0
google-api-python-client>=2.116.0
google-auth-oauthlib>=1.2.0
google-auth>=2.25.0
PyYAML>=6.0.1